# docs/_static/files/swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "paths": {
...
        "/api/browse/{vendor}": {
            "parameters": [
                {
                    "name": "vendor",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Get the known products for a vendor",
                "operationId": "get_vendor_products",
                "tags": [
                    "api"
                ]
            }
        },
...
        "/api/cve/{vulnerability_id}": {
            "parameters": [
                {
                    "name": "vulnerability_id",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "delete": {
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "403": {
                        "description": "Admin permission required."
                    },
                    "422": {
                        "description": "Not possible to edit a vulnerability from the requested source."
                    }
                },
                "summary": "Endpoint for deleting a vulnerability",
                "description": "Delete a vulnerability from the local source.\nWe only accept to delete vulnerabilities from the local source.",
                "operationId": "delete_vulnerability",
                "tags": [
                    "api"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Get a vulnerability.",
                "operationId": "get_vulnerability",
                "tags": [
                    "api"
                ]
            }
        },
...
        "/api/search/{vendor}/{product}": {
            "parameters": [
                {
                    "name": "vendor",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "product",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Get the the vulnerabilities per vendor and a specific product",
                "operationId": "get_vendor_product_vulnerabilities",
                "tags": [
                    "api"
                ]
            }
        },
...
        "/last/{number}": {
            "parameters": [
                {
                    "name": "number",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Get the last CVEs",
                "operationId": "get_last",
                "tags": [
                    "default"
                ]
            }
        },
        "/last/{source}": {
            "parameters": [
                {
                    "name": "source",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Get the last CVEs",
                "operationId": "get_last",
                "tags": [
                    "default"
                ]
            }
        },
        "/last/{source}/{number}": {
            "parameters": [
                {
                    "name": "source",
                    "in": "path",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "number",
                    "in": "path",
                    "required": true,
                    "type": "integer"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Get the last CVEs",
                "operationId": "get_last",
                "tags": [
                    "default"
                ]
            }
        },
...
        "/vulnerability/": {
            "post": {
                "responses": {
                    "200": {
                        "description": "Success."
                    },
                    "400": {
                        "description": "JSON validation failed."
                    },
                    "403": {
                        "description": "Reporter permission required."
                    },
                    "422": {
                        "description": "Not possible to edit a vulnerability from the requested source."
                    }
                },
                "summary": "Endpoint for creating and editing vulnerabilities in the local source",
                "description": "Create a vulnerability with the CVE version 5 format.",
                "operationId": "post_vulnerabilities_list",
                "parameters": [
                    {
                        "name": "payload",
                        "required": true,
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "data": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "tags": [
                    "default"
                ]
            }
        },
        "/vulnerability/{vulnerability_id}": {
            "parameters": [
                {
                    "name": "vulnerability_id",
                    "in": "path",
                    "required": true,
                    "type": "string"
                }
            ],
            "delete": {
                "responses": {
                    "204": {
                        "description": "Success."
                    },
                    "403": {
                        "description": "Admin permission required."
                    },
                    "422": {
                        "description": "Not possible to edit a vulnerability from the requested source."
                    }
                },
                "summary": "Endpoint for deleting a vulnerability",
                "description": "Delete a vulnerability from the local source.\nWe only accept to delete vulnerabilities from the local source.",
                "operationId": "delete_vulnerability",
                "tags": [
                    "default"
                ]
            },
            "get": {
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "description": "Get a vulnerability.",
                "operationId": "get_vulnerability",
                "tags": [
                    "default"
                ]
            }
        }
    },
    "info": {
        "title": "Vulnerability Lookup API",
        "version": "1.7.0",
        "description": "<a href='https://www.circl.lu' rel='noreferrer' target='_blank'><img src='https://www.circl.lu/assets/images/circl-logo.png' /></a><br />API to query <a href='https://github.com/cve-search/vulnerability-lookup' rel='noreferrer' target='_blank'>Vulnerability Lookup</a>.<br /><br />Back to the <a href='/'>main page</a>.",
        "license": {
            "name": "GNU Affero General Public License version 3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.html"
        }
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "securityDefinitions": {
        "apikey": {
            "type": "apiKey",
            "in": "header",
            "name": "X-API-KEY"
        }
    },
    "security": [
        {
            "apikey": []
        }
    ],
...
}